requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docopt==0.6.2
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.1
requests==2.31.0
rq==1.15.1
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
times==0.7
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.1
yarg==0.1.9